renamed some files to snake case
