+++ lib/routing/go_router.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:youdoyou/common_widgets/error_widget.dart';
 import 'package:youdoyou/features/todos/domain/todo_model.dart';
-import 'package:youdoyou/features/todos/presentation/detailScreen/detail_screen.dart';
+import 'package:youdoyou/features/todos/presentation/detail_screen/detail_screen.dart';
 import 'package:youdoyou/features/authentication/presentation/auth_page.dart';
 import 'package:youdoyou/features/todos/presentation/home_screen/home.dart';
 import 'package:youdoyou/routing/routes.dart';
